Renamed file  to , as described in point 6 of the following [pull request](https://github.com/Baekalfen/PyBoy/pull/22)
